protificando tabela
