+++ src/components/App/App.tsx
@@ -1,9 +1,16 @@
 
 import Container from '../../shared/Container';
-import Table from '../../shared/Table';
+import Table, { TableHeader } from '../../shared/Table';
+import Products from '../../shared/Table/Table.mockdata';
 import Header from '../Header';
 import './App.css';
 
+const headers: TableHeader[] = [
+  { key: 'id', value: '#' },
+  { key: 'name', value: 'Product' },
+  { key: 'price', value: 'Price', right: true },
+  { key: 'stock', value: 'Available Stock', right: true }
+]
 
 function App() {
 
@@ -12,11 +19,11 @@ function App() {
       <Header title="AlgaStock" />
 
       <Container>
-       <>
-        <Table />
-       </>
+        <Table
+          headers={headers}
+          data={Products}
+        />
       </Container>
-
     </div>
   );
 }
