+++ src/shared/Table/index.ts
@@ -1,1 +1,3 @@
-export { default } from './Table'   
+
+export { default } from './Table'
+export * from './Table'
